+++ Classes/TellTimeViewController.m
@@ -227,7 +227,7 @@
 		else [choices setTitle:[NSString stringWithFormat:@"%i:%02i", h, m] forSegmentAtIndex:i];}
 	//load answer in a random index, but remember where it is
 	randomNumber.rangeLow = 0;
-	randomNumber.rangeHigh = choices.numberOfSegments-1;
+	randomNumber.rangeHigh = (int) choices.numberOfSegments-1;
 	self.answerIndex = randomNumber.randomInteger;
 	if (((self.activityLevel == kActLevelYellowBelt) || (self.activityLevel == kActLevelGreenBelt)) && (answerMinutes == 0)) [choices setTitle:[NSString stringWithFormat:@"%i o'clock", answerHours] forSegmentAtIndex:self.answerIndex];
 		else [choices setTitle:[NSString stringWithFormat:@"%i:%02i", answerHours, answerMinutes] forSegmentAtIndex:self.answerIndex];
